# docs/swagger.yaml
basePath: /api
definitions:
  presenter.GetUserByIdOrNameResponse:
    properties:
      id:
        type: string
      limit:
        type: integer
      username:
        type: string
    type: object
  presenter.SignInInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  presenter.SignInResponse:
    properties:
      token:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
  presenter.SignUpInput:
    properties:
      limit:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
  presenter.SignUpResponse:
    properties:
      id:
        type: string
      limit:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: mailto:thuongtruongofficial@gmail.com
    name: Tran Nguyen Thuong Truong
    url: https://github.com/thuongtruong1009/zoomer
  description: Enter the token with the `Bearer ` prefix, e.g. `Bearer jwt_token_string`.
  license:
    name: Apache 2.0
    url: https://github.com/thuongtruong1009/zoomer/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Zoomer
  version: "1.1"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Login to user account
      parameters:
      - description: Login user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/presenter.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.SignInResponse'
        "400":
          description: Bad Request
          schema:
            type: error
        "500":
          description: Internal Server Error
          schema:
            type: error
      summary: Login to user account
      tags:
      - auth
  /auth/signout:
    post:
      consumes:
      - application/json
      description: Logout user credentials
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: error
        "401":
          description: Unauthorized
          schema:
            type: error
        "500":
          description: Internal Server Error
          schema:
            type: error
      security:
      - bearerAuth: []
      summary: Logout user credentials
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: Create User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/presenter.SignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presenter.SignUpResponse'
        "400":
          description: Bad Request
          schema:
            type: error
        "500":
          description: Internal Server Error
          schema:
            type: error
      summary: Create a new user
      tags:
      - auth
  /users/{idOrName}:
    get:
      consumes:
      - application/json
      description: Get user by id or name
      parameters:
      - description: id or name of user
        in: path
        name: idOrName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.GetUserByIdOrNameResponse'
        "401":
          description: Unauthorized
          schema:
            type: error
        "500":
          description: Internal Server Error
          schema:
            type: error
      summary: Get user by id or name
      tags:
      - users
schemes:
- http
securityDefinitions:
  XFirebaseBearer:
    description: Enter the token with the `Bearer ` prefix, e.g. `Bearer jwt_token_string`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"